django/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-ckeditor==6.3.2
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-faicon==0.1.0
django-filter==23.5
django-jazzmin==2.4.4
django-js-asset==2.2.0
django-tinymce==3.6.1
django-translation==1.2
django-translations==1.3.0
django-user-agents==0.4.0
djangorestframework==3.12.2
drf-yasg==1.21.4
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
Pillow==9.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sqlparse==0.5.1
typing_extensions==4.12.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
gunicorn==20.1.0